refactor: controller.js 함수 별로 분리
