--- js/controller.js
@@ -1,57 +1,0 @@
-import { gameInformationStore } from "./model.js";
-import { raceTimeTemplate } from "./view/raceTimeTemplate.js";
-import { gameProcesstemplate } from "./view/gameProcessTemplate.js";
-import { raceTimeEventListener } from "./init.js";
-
-const carNameInput = document.querySelector("#carNameInput");
-const userInputComponent = document.querySelector("#user-input-component");
-const carNameInputButton = document.querySelector("#carNameInputSubmitButton");
-
-export const handleCarNameInput = (event) => {
-  event.preventDefault();
-
-  let carNameInputValue = carNameInput.value.split(",");
-  carNameInputValue = carNameInputValue.map((item) => {
-    return item.trim();
-  });
-
-  for (let item of carNameInputValue) {
-    if (item.length > 5) {
-      alert("5자 이하의 자동차 이름을 입력해주세요.");
-      document.querySelector("#carNameInput").value = "";
-      return;
-    }
-  }
-
-  gameInformationStore.carName = carNameInputValue;
-  carNameInput.disabled = true;
-  carNameInputButton.disabled = true;
-
-  const newSection = document.createElement("section");
-  newSection.innerHTML = raceTimeTemplate;
-  userInputComponent.appendChild(newSection);
-
-  raceTimeEventListener();
-};
-
-export const handleRaceTimeInput = (event) => {
-  event.preventDefault();
-  const raceTimeInput = document.querySelector("#raceTimeInput");
-  const raceTimeSubmitButton = document.querySelector("#raceTimeSubmitButton");
-
-  gameInformationStore.raceTime = Number(raceTimeInput.value);
-
-  raceTimeInput.disabled = true;
-  raceTimeSubmitButton.disabled = true;
-
-  carNameRender();
-};
-
-const carNameRender = () => {
-  const divApp = document.querySelector("#app");
-
-  const newDiv = document.createElement("div");
-
-  newDiv.innerHTML = gameProcesstemplate(gameInformationStore.carName);
-  divApp.appendChild(newDiv);
-};
